Move bg bin utility functions from trends util to bloodglucose util
